requirements.txt
alembic==1.12.1
amqp==5.2.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
paramiko==3.4.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.1
XlsxWriter==3.1.9
